Changes intended for usergroup-calculation-into-class
The purpose of error-logging was to create an error logging class capable
to log error into a file. Errors preventing me to actually run the plugin
and test the class led me to fix what should have been done on the other
branch. This commit is intended to be merged to
usergroup-calculation-into-class.